MovingPath hierarchy documentation
Methods documented for the interface MovingPath and its implementing
classes Path and CompoundPath. The unused AbstractPath class was
removed. Alternative constructors for Path and CompoundPath added that
begin at the origin (0, 0). Renamed xWithRespectToOrigin() and
yWithRespectToOrigin() to getX() and getY() respectively. Instance
variables for Path and CompoundPath made final and renamed to fit
convention. Removed unused imports.