wrapMessage revised
Documented and revised TextHandler.wrapMessage. New FACE_WRAP size added to TextHandler and referenced in DialogueBox. The example message is shown at every step: "* Do you need some help..? * Press the switch on the wall." The method wraps the message around "* Press". It determines when to indent by an asterisk.